pipelines/pyproject.toml
[tool.poetry]
name = "pipelines"
version = "0.1.0"
description = "E2E Vertex AI Pipeline"
authors = ["Example User <user@example.com>"]
readme = "README.md"
packages = [
    { include = "pipelines", from = "src" },
]

[tool.poetry.dependencies]
python = "3.10.14"
kfp = "^2.7.0"
google-cloud-aiplatform = "^1.55.0"
jinja2 = "^3.1.4"
tensorflow = "^2.16.1"
google-cloud-pipeline-components = "^2.14.1"
components = { path = "../components", develop = true }

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
